+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.0 under development
 
 - New #137: Add `$attributes` parameter to `Html::img()` (@alien-art)
+- New #150: Add class for tag `small` and method `Html::small()` (@dood-)
 
 ## 3.0.0 November 06, 2022